+++ resources/views/adminarea/pages/admin-attributes.blade.php
@@ -39,7 +39,7 @@
                                         {{--{{ Form::button(trans('cortex/auth::common.submit'), ['class' => 'btn btn-primary btn-flat', 'type' => 'submit']) }}--}}
                                     {{--</div>--}}
 
-                                    {{--@include('cortex/foundation::common.partials.timestamps', ['model' => $admin])--}}
+                                    {{--@include('cortex/foundation::adminarea.partials.timestamps', ['model' => $admin])--}}
 
                                 {{--</div>--}}