+++ src/Http/Requests/Adminarea/AccountTwoFactorPhoneRequest.php
@@ -11,7 +11,7 @@ class AccountTwoFactorPhoneRequest extends FormRequest
     /**
      * Determine if the user is authorized to make this request.
      *
-     * @throws \Cortex\Foundation\Exceptions\GenericException
+     * @throws \Cortex\Auth\Exceptions\AccountException
      *
      * @return bool
      */